requirements.txt
disnake[voice]
aiofiles
pillow
python-decouple
requests
pyrebase4
